requirements.txt
Flask
pandas
numpy
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
APScheduler
python-dotenv